merit-travelops-demo/pyproject.toml
[project]
name = "travelops-merit-demo"
version = "0.1.0"
description = "Merit demo repository showcasing fix-oriented failure clustering for TravelOps Assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "appmerit>=0.1.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "tests"]

[tool.merit]
test-paths = ["tests"]
concurrency = 4

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.uv.sources]
appmerit = { git = "https://github.com/appMerit/merit", branch = "main" }
